# Refactor callable lookup