--- a/lib/AST/Type.cpp
+++ b/lib/AST/Type.cpp
@@ -2006,7 +2006,7 @@ bool TypeBase::satisfiesClassConstraint() {
   return mayHaveSuperclass() || isObjCExistentialType();
 }
 
-bool TypeBase::isCallableNominalType(DeclContext *dc) {
+bool TypeBase::isCallAsFunctionType(DeclContext *dc) {
   // Don't allow callAsFunction to be used with dynamic lookup.
   if (isAnyObject())
     return false;