--- a/lib/Sema/CSSimplify.cpp
+++ b/lib/Sema/CSSimplify.cpp
@@ -1185,7 +1185,7 @@ public:
         argIdx == *UnlabeledTrailingClosureArgIndex && Callee) {
       if (auto *ctor = dyn_cast<ConstructorDecl>(Callee.get())) {
         auto resultTy = ctor->getResultInterfaceType();
-        if (resultTy->isCallableNominalType(CS.DC))
+        if (resultTy->isCallAsFunctionType(CS.DC))
           return true;
       }
     }
@@ -10985,7 +10985,7 @@ bool ConstraintSystem::simplifyAppliedOverloadsImpl(
         auto *dc = decl->getDeclContext();
         if (auto *parent = dc->getSelfNominalTypeDecl()) {
           auto type = parent->getDeclaredInterfaceType();
-          if (type->isCallableNominalType(DC))
+          if (type->isCallAsFunctionType(DC))
             return false;
         }
       }
@@ -11330,7 +11330,7 @@ ConstraintSystem::simplifyApplicableFnConstraint(
   // Handle applications of types with `callAsFunction` methods.
   // Do this before stripping optional types below, when `shouldAttemptFixes()`
   // is true.
-  if (desugar2->isCallableNominalType(DC)) {
+  if (desugar2->isCallAsFunctionType(DC)) {
     auto memberLoc = getConstraintLocator(
         locator.withPathElement(ConstraintLocator::ImplicitCallAsFunction));
     // Add a `callAsFunction` member constraint, binding the member type to a
@@ -11391,7 +11391,7 @@ ConstraintSystem::simplifyApplicableFnConstraint(
       // trailing closure(s), closure(s) might not belong to
       // the constructor but rather to implicit `callAsFunction`,
       // there is no way to determine that without trying.
-      if (resultTy->isCallableNominalType(DC) &&
+      if (resultTy->isCallAsFunctionType(DC) &&
           argumentList->hasAnyTrailingClosures()) {
         auto *calleeLoc = getCalleeLocator(argumentsLoc);
 
@@ -11553,7 +11553,7 @@ ConstraintSystem::simplifyApplicableFnConstraint(
     // once constraint generation is done.
     if (getPhase() == ConstraintSystemPhase::ConstraintGeneration &&
         argumentList->hasAnyTrailingClosures() &&
-        instance2->isCallableNominalType(DC)) {
+        instance2->isCallAsFunctionType(DC)) {
       return formUnsolved(/*activate=*/true);
     }