Move off `Type` based requests and onto `Decl` based requests, utilizing name lookup's `extractDirectlyReferencedNominalTypes` utility. This allows us to better cache the results, and avoids the need to guard against type variable inputs when deciding whether or not to cache.