@hamishknight 2022-07-22
@swift-ci please test compiler performance