@hamishknight 2022-07-23
@swift-ci please test compiler performance