@hamishknight 2022-07-23
# Summary for main full

**Regressions found (see below)**

- [Debug-batch](#debug-batch-3535b3c9009a87b)
    - [debug-batch brief](#debug-batch-brief-3535b3c9009a87b)
    - [debug-batch detailed](#debug-batch-detailed-3535b3c9009a87b)
- [Release](#release-3535b3c9009a87b)
    - [release brief](#release-brief-3535b3c9009a87b)
    - [release detailed](#release-detailed-3535b3c9009a87b)



<a name="debug-batch-3535b3c9009a87b">

# Debug-batch

<a name="debug-batch-brief-3535b3c9009a87b">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 73,911,162,135,216 | 73,924,044,539,703 | 12,882,404,487 | 0.02%
LLVM.NumLLVMBytesOutput | 1,837,804,914 | 1,837,851,876 | 46,962 | 0.0%
</details>

<a name="debug-batch-detailed-3535b3c9009a87b">

## debug-batch detailed
<details>
<summary>Regressed (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Sema.HasDynamicCallableAttributeRequest | 0 | 17,106 | 17,106 | 100.0% :no_entry:
</details>
<details>
<summary>Improved (4)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumProcessFailures | 41 | 40 | -1 | -2.44% :white_check_mark:
Sema.HasDynamicMemberLookupAttributeRequest | 625,857 | 485,036 | -140,820 | -22.5% :white_check_mark:
Sema.IsCallableNominalTypeRequest | 10,675 | 0 | -10,675 | -100.0% :white_check_mark:
Sema.SuperclassTypeRequest | 68,215 | 65,260 | -2,954 | -4.33% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (279)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 2,825,349 | 2,825,352 | 3 | 0.0%
AST.ImportSetCacheMiss | 625,771 | 625,790 | 18 | 0.0%
AST.ImportSetFoldHit | 319,024 | 319,045 | 21 | 0.01%
AST.ImportSetFoldMiss | 306,747 | 306,745 | -2 | -0.0%
AST.ModuleShadowCacheHit | 5,323 | 5,323 | 0 | 0.0%
AST.ModuleShadowCacheMiss | 2,352 | 2,352 | 0 | 0.0%
AST.ModuleVisibilityCacheHit | 12,754,587 | 12,754,433 | -154 | -0.0%
AST.ModuleVisibilityCacheMiss | 119,111 | 119,105 | -6 | -0.01%
AST.NumASTBytesAllocated | 43,737,820,959 | 43,763,253,701 | 25,432,742 | 0.06%
AST.NumASTScopeExpansions | 6,896,151 | 6,895,981 | -169 | -0.0%
AST.NumASTScopeLookups | 7,308,142 | 7,307,894 | -248 | -0.0%
AST.NumDecls | 289,929 | 289,924 | -5 | -0.0%
AST.NumDependencies | 334,602 | 334,599 | -2 | -0.0%
AST.NumIncrementalDependencies | 12,455 | 12,455 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 241,518 | 241,518 | 0 | 0.0%
AST.NumLocalTypeDecls | 350 | 350 | 0 | 0.0%
AST.NumModuleLookupClassMember | 8,310 | 8,310 | 0 | 0.0%
AST.NumModuleLookupValue | 74,946,243 | 74,944,653 | -1,590 | -0.0%
AST.NumObjCMethods | 52,876 | 52,876 | 0 | 0.0%
AST.NumOperators | 1,313 | 1,313 | 0 | 0.0%
AST.NumPrecedenceGroups | 243 | 243 | 0 | 0.0%
AST.NumReferencedDynamicNames | 204 | 204 | 0 | 0.0%
AST.NumReferencedMemberNames | 11,826,102 | 11,825,877 | -225 | -0.0%
AST.NumReferencedTopLevelNames | 1,120,468 | 1,120,455 | -13 | -0.0%
AST.NumSourceBuffers | 333,131 | 333,127 | -4 | -0.0%
AST.NumSourceLines | 9,578,512 | 9,578,392 | -120 | -0.0%
AST.NumSourceLinesPerSecond | 7,319,622 | 7,357,383 | 37,761 | 0.52%
AST.NumTotalClangImportedEntities | 1,742,379 | 1,743,511 | 1,131 | 0.06%
Frontend.MaxMallocUsage | 717,995,819,960 | 718,491,975,904 | 496,155,944 | 0.07%
Frontend.NumInstructionsExecuted | 73,911,162,135,216 | 73,924,044,539,703 | 12,882,404,487 | 0.02%
IRGen.IRGenRequest | 25,934 | 25,935 | 1 | 0.0%
IRGen.OptimizedIRRequest | 0 | 0 | 0 | 0.0%
IRGen.SymbolObjectCodeRequest | 0 | 0 | 0 | 0.0%
IRModule.NumGOTEntries | 242,605 | 242,613 | 7 | 0.0%
IRModule.NumIRAliases | 203,735 | 203,738 | 2 | 0.0%
IRModule.NumIRBasicBlocks | 7,753,674 | 7,753,839 | 165 | 0.0%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 3,503,895 | 3,503,979 | 83 | 0.0%
IRModule.NumIRGlobals | 3,730,497 | 3,730,567 | 70 | 0.0%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 80,442,759 | 80,444,460 | 1,701 | 0.0%
IRModule.NumIRNamedMetaData | 128,862 | 128,867 | 5 | 0.0%
IRModule.NumIRValueSymbols | 7,283,332 | 7,283,485 | 152 | 0.0%
LLVM.NumLLVMBytesOutput | 1,837,804,914 | 1,837,851,876 | 46,962 | 0.0%
Parse.CodeCompletionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 306,639 | 306,639 | 0 | -0.0%
Parse.NumIterableDeclContextParsed | 657,345 | 657,335 | -10 | -0.0%
Parse.ParseAbstractFunctionBodyRequest | 537,754 | 537,757 | 3 | 0.0%
Parse.ParseMembersRequest | 515,733 | 515,720 | -13 | -0.0%
Parse.ParseSourceFileRequest | 328,080 | 328,080 | 0 | 0.0%
SILGen.ASTLoweringRequest | 27,064 | 27,065 | 1 | 0.0%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 2,572,563 | 2,572,618 | 55 | 0.0%
SILModule.NumSILGenGlobalVariables | 119,225 | 119,226 | 1 | 0.0%
SILModule.NumSILGenVtables | 33,492 | 33,491 | 0 | -0.0%
SILModule.NumSILGenWitnessTables | 139,565 | 139,568 | 2 | 0.0%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 2,962,671 | 2,962,774 | 103 | 0.0%
SILModule.NumSILOptGlobalVariables | 120,993 | 120,994 | 1 | 0.0%
SILModule.NumSILOptVtables | 33,717 | 33,716 | 0 | -0.0%
SILModule.NumSILOptWitnessTables | 171,746 | 171,751 | 5 | 0.0%
SILOptimizer.ExecuteSILPipelineRequest | 106,032 | 106,036 | 4 | 0.0%
SILOptimizer.LoweredSILRequest | 0 | 0 | 0 | 0.0%
Sema.ABIMembersRequest | 193,756 | 193,759 | 3 | 0.0%
Sema.AbstractGenericSignatureRequest | 124,944 | 124,999 | 54 | 0.04%
Sema.AccessLevelRequest | 12,689,335 | 12,699,297 | 9,962 | 0.08%
Sema.ActorIsolationRequest | 5,630,768 | 5,630,631 | -136 | -0.0%
Sema.AllMembersRequest | 92,367 | 92,367 | 0 | 0.0%
Sema.AnyObjectLookupRequest | 299 | 299 | 0 | 0.0%
Sema.ApplyAccessNoteRequest | 3,563,897 | 3,564,063 | 166 | 0.0%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 36,801 | 36,802 | 0 | 0.0%
Sema.AttachedPropertyWrapperTypeRequest | 19,165 | 19,166 | 0 | 0.0%
Sema.AttachedPropertyWrappersRequest | 1,672,176 | 1,672,139 | -37 | -0.0%
Sema.AttachedResultBuilderRequest | 838,490 | 838,490 | 0 | 0.0%
Sema.BodyInitKindRequest | 60,510 | 60,510 | 0 | 0.0%
Sema.CXXNamespaceMemberLookup | 0 | 0 | 0 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 126,799 | 126,798 | -1 | -0.0%
Sema.CheckDistributedFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 9,166 | 9,166 | 0 | 0.0%
Sema.CheckRedeclarationRequest | 1,859,766 | 1,859,759 | -7 | -0.0%
Sema.ClangDirectLookupRequest | 14,608 | 14,608 | 0 | 0.0%
Sema.ClangRecordMemberLookup | 14,608 | 14,608 | 0 | 0.0%
Sema.ClassAncestryFlagsRequest | 131,555 | 131,548 | -7 | -0.01%
Sema.ClosureEffectsRequest | 103,348 | 103,334 | -14 | -0.01%
Sema.ClosureHasExplicitResultRequest | 35,963 | 35,962 | -1 | -0.0%
Sema.CodeCompletionFileRequest | 0 | 0 | 0 | 0.0%
Sema.CollectOverriddenDeclsRequest | 9,528,945 | 9,544,402 | 15,457 | 0.16%
Sema.CompareDeclSpecializationRequest | 609,536 | 609,664 | 128 | 0.02%
Sema.ConditionalRequirementsRequest | 650,345 | 650,629 | 283 | 0.04%
Sema.ConformanceHasEffectRequest | 0 | 0 | 0 | 0.0%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 22,911 | 22,911 | 0 | 0.0%
Sema.CustomAttrTypeRequest | 938 | 938 | 0 | 0.0%
Sema.CxxRecordSemantics | 0 | 0 | 0 | 0.0%
Sema.DefaultAndMaxAccessLevelRequest | 103,670 | 103,673 | 3 | 0.0%
Sema.DefaultArgumentExprRequest | 87,883 | 87,869 | -14 | -0.02%
Sema.DefaultArgumentInitContextRequest | 694 | 694 | 0 | 0.0%
Sema.DefaultArgumentTypeRequest | 19,450 | 19,450 | 0 | -0.0%
Sema.DefaultDefinitionTypeRequest | 7,538 | 7,537 | 0 | -0.01%
Sema.DefaultTypeRequest | 335,697 | 335,706 | 9 | 0.0%
Sema.DerivativeAttrOriginalDeclRequest | 40 | 40 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 53,678,091 | 53,646,274 | -31,817 | -0.06%
Sema.DirectOperatorLookupRequest | 1,884,327 | 1,884,299 | -28 | -0.0%
Sema.DirectPrecedenceGroupLookupRequest | 764,284 | 764,296 | 11 | 0.0%
Sema.DistributedModuleIsAvailableRequest | 0 | 0 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 6,008,235 | 6,008,109 | -126 | -0.0%
Sema.EnumRawTypeRequest | 34,824 | 34,824 | 0 | 0.0%
Sema.EnumRawValuesRequest | 19,066 | 19,066 | 0 | 0.0%
Sema.ExistentialConformsToSelfRequest | 21,773 | 21,777 | 4 | 0.02%
Sema.ExistentialRequiresAnyRequest | 16,988 | 16,988 | 0 | 0.0%
Sema.ExtendedNominalRequest | 575,629 | 575,629 | 0 | 0.0%
Sema.ExtendedTypeRequest | 180,752 | 180,754 | 2 | 0.0%
Sema.FragileFunctionKindRequest | 2,099,717 | 2,099,694 | -22 | -0.0%
Sema.FunctionOperatorRequest | 18,829 | 18,830 | 1 | 0.01%
Sema.GenericParamListRequest | 6,059,338 | 6,060,249 | 911 | 0.02%
Sema.GenericSignatureRequest | 3,059,113 | 3,059,808 | 695 | 0.02%
Sema.GetDestructorRequest | 40,675 | 40,673 | -2 | -0.0%
Sema.GetDistributedActorArgumentDecodingMethodRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedActorIDPropertyRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedActorImplicitCodableRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedActorInvocationDecoderRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedActorSystemPropertyRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedActorSystemRemoteCallFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedRemoteCallArgumentInitFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedRemoteCallTargetInitFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedTargetInvocationDecoderDecodeNextArgumentFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordArgumentFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordErrorTypeFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordReturnTypeFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedTargetInvocationResultHandlerOnReturnFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedThunkRequest | 0 | 0 | 0 | 0.0%
Sema.GetImplicitSendableRequest | 298,778 | 298,541 | -237 | -0.08%
Sema.GetSourceFileAsyncNode | 99 | 99 | 0 | 0.0%
Sema.GlobalActorAttributeRequest | 6,620,627 | 6,620,628 | 1 | 0.0%
Sema.GlobalActorInstanceRequest | 980 | 980 | 0 | 0.0%
Sema.HasCircularInheritedProtocolsRequest | 18,032 | 18,032 | 0 | 0.0%
Sema.HasCircularRawValueRequest | 18,633 | 18,633 | 0 | 0.0%
Sema.HasDefaultInitRequest | 83,403 | 83,401 | -1 | -0.0%
Sema.HasImplementationOnlyImportsRequest | 325,130 | 325,130 | 0 | 0.0%
Sema.HasIsolatedSelfRequest | 11,969,850 | 11,975,152 | 5,301 | 0.04%
Sema.HasMemberwiseInitRequest | 35,487 | 35,487 | 0 | 0.0%
Sema.HasMissingDesignatedInitializersRequest | 24,691 | 24,691 | 0 | 0.0%
Sema.HasUserDefinedDesignatedInitRequest | 83,415 | 83,413 | -1 | -0.0%
Sema.InferredGenericSignatureRequest | 226,893 | 226,881 | -12 | -0.01%
Sema.InheritedDeclsReferencedRequest | 6,438,726 | 6,448,623 | 9,897 | 0.15%
Sema.InheritedProtocolsRequest | 707,736 | 708,120 | 384 | 0.05%
Sema.InheritedTypeRequest | 418,009 | 418,032 | 23 | 0.01%
Sema.InheritsSuperclassInitializersRequest | 38,057 | 38,057 | 0 | 0.0%
Sema.InitKindRequest | 151,964 | 151,963 | -1 | -0.0%
Sema.InterfaceTypeRequest | 17,128,316 | 17,143,385 | 15,068 | 0.09%
Sema.IsABICompatibleOverrideRequest | 275,444 | 275,450 | 6 | 0.0%
Sema.IsAccessorTransparentRequest | 524,791 | 524,797 | 6 | 0.0%
Sema.IsActorRequest | 1,901,970 | 1,902,322 | 352 | 0.02%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDeclRefinementOfRequest | 53,009 | 53,012 | 3 | 0.01%
Sema.IsDefaultActorRequest | 54,499 | 54,498 | 0 | -0.0%
Sema.IsDistributedActorRequest | 2,110,219 | 2,110,788 | 569 | 0.03%
Sema.IsDynamicRequest | 2,614,527 | 2,614,566 | 39 | 0.0%
Sema.IsFinalRequest | 3,797,654 | 3,800,234 | 2,580 | 0.07%
Sema.IsGetterMutatingRequest | 653,172 | 653,163 | -9 | -0.0%
Sema.IsImplicitlyUnwrappedOptionalRequest | 3,308,691 | 3,309,115 | 423 | 0.01%
Sema.IsObjCRequest | 2,459,886 | 2,460,087 | 200 | 0.01%
Sema.IsSafeUseOfCxxDecl | 0 | 0 | 0 | 0.0%
Sema.IsSetterMutatingRequest | 564,289 | 564,280 | -9 | -0.0%
Sema.IsStaticRequest | 1,431,828 | 1,431,851 | 23 | 0.0%
Sema.LazyStoragePropertyRequest | 3,067 | 3,067 | 0 | 0.0%
Sema.LookupAllConformancesInContextRequest | 2,042,969 | 2,045,171 | 2,201 | 0.11%
Sema.LookupConformanceInModuleRequest | 65,999,557 | 66,015,480 | 15,923 | 0.02%
Sema.LookupInModuleRequest | 7,570,364 | 7,572,115 | 1,751 | 0.02%
Sema.LookupInfixOperatorRequest | 67,136 | 67,141 | 5 | 0.01%
Sema.LookupPostfixOperatorRequest | 82 | 82 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 30,056 | 30,057 | 1 | 0.0%
Sema.LookupPrefixOperatorRequest | 392 | 392 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 64 | 64 | 0 | 0.0%
Sema.ModuleImplicitImportsRequest | 9,166 | 9,166 | 0 | 0.0%
Sema.ModuleLibraryLevelRequest | 0 | 0 | 0 | 0.0%
Sema.ModuleQualifiedLookupRequest | 2,855,617 | 2,857,487 | 1,870 | 0.07%
Sema.NamedLazyMemberLoadSuccessCount | 32,746,132 | 32,752,393 | 6,260 | 0.02%
Sema.NamingPatternRequest | 240,503 | 240,503 | 0 | 0.0%
Sema.NeedsNewVTableEntryRequest | 929,191 | 929,186 | -4 | -0.0%
Sema.NumAccessorBodiesSynthesized | 414,015 | 414,018 | 3 | 0.0%
Sema.NumAccessorsSynthesized | 519,947 | 519,949 | 2 | 0.0%
Sema.NumConformanceAccessPathsRecorded | 700,847 | 701,185 | 337 | 0.05%
Sema.NumConformancesDeserialized | 4,457,402 | 4,462,693 | 5,290 | 0.12%
Sema.NumConstraintScopes | 45,417,429 | 45,404,931 | -12,498 | -0.03%
Sema.NumConstraintsConsideredForEdgeContraction | 1,625,918 | 1,625,890 | -28 | -0.0%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 0 | 0 | 0 | 0.0%
Sema.NumDeclsDeserialized | 73,625,650 | 73,696,925 | 71,275 | 0.1%
Sema.NumDeclsTypechecked | 2,625,241 | 2,625,217 | -24 | -0.0%
Sema.NumGenericSignatureBuilders | 0 | 0 | 0 | 0.0%
Sema.NumLazyIterableDeclContexts | 10,926,007 | 10,931,927 | 5,920 | 0.05%
Sema.NumLazyRequirementSignatures | 899,178 | 899,505 | 327 | 0.04%
Sema.NumLazyRequirementSignaturesLoaded | 622,440 | 622,743 | 303 | 0.05%
Sema.NumLeafScopes | 34,268,622 | 34,259,632 | -8,989 | -0.03%
Sema.NumRedundantRequirementSteps | 0 | 0 | 0 | 0.0%
Sema.NumRequirementMachineCompletionSteps | 21,107,650 | 21,118,006 | 10,356 | 0.05%
Sema.NumRequirementMachineUnifiedConcreteTerms | 904,236 | 906,335 | 2,099 | 0.23%
Sema.NumRequirementMachines | 1,790,826 | 1,791,488 | 662 | 0.04%
Sema.NumTypesDeserialized | 21,081,664 | 21,094,519 | 12,855 | 0.06%
Sema.NumUnloadedLazyIterableDeclContexts | 7,936,655 | 7,936,852 | 197 | 0.0%
Sema.OpaqueReadOwnershipRequest | 456,186 | 456,189 | 3 | 0.0%
Sema.OpaqueResultTypeRequest | 44 | 44 | 0 | 0.0%
Sema.OperatorPrecedenceGroupRequest | 1,325 | 1,325 | 0 | 0.0%
Sema.OverriddenDeclsRequest | 5,015,012 | 5,017,446 | 2,434 | 0.05%
Sema.ParamSpecifierRequest | 1,920,445 | 1,920,426 | -19 | -0.0%
Sema.PatternBindingEntryRequest | 650,880 | 650,870 | -10 | -0.0%
Sema.PatternTypeRequest | 754,704 | 754,687 | -17 | -0.0%
Sema.PolymorphicEffectKindRequest | 106,760 | 106,758 | -2 | -0.0%
Sema.PolymorphicEffectRequirementsRequest | 0 | 0 | 0 | 0.0%
Sema.PreCheckResultBuilderRequest | 32 | 32 | 0 | 0.0%
Sema.PrimaryAssociatedTypesRequest | 7,471 | 7,471 | 0 | 0.0%
Sema.PrimarySourceFilesRequest | 9,166 | 9,166 | 0 | 0.0%
Sema.PropertyWrapperAuxiliaryVariablesRequest | 528,425 | 528,416 | -9 | -0.0%
Sema.PropertyWrapperBackingPropertyTypeRequest | 19,165 | 19,166 | 0 | 0.0%
Sema.PropertyWrapperInitializerInfoRequest | 537,347 | 537,338 | -9 | -0.0%
Sema.PropertyWrapperLValuenessRequest | 648 | 648 | 0 | 0.0%
Sema.PropertyWrapperMutabilityRequest | 648,307 | 648,298 | -9 | -0.0%
Sema.PropertyWrapperTypeInfoRequest | 264 | 264 | 0 | 0.0%
Sema.ProtocolDependenciesRequest | 538,353 | 538,772 | 418 | 0.08%
Sema.ProtocolRequiresClassRequest | 46,829 | 46,915 | 86 | 0.18%
Sema.ProvideDefaultImplForRequest | 9,528,945 | 9,544,402 | 15,457 | 0.16%
Sema.QualifiedLookupRequest | 7,445,381 | 7,438,652 | -6,729 | -0.09%
Sema.RangeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.RenamedDeclRequest | 170 | 170 | 0 | 0.0%
Sema.RequirementRequest | 180,567 | 180,566 | -1 | -0.0%
Sema.RequirementSignatureRequest | 680,357 | 680,730 | 373 | 0.05%
Sema.RequiresOpaqueAccessorsRequest | 1,634,570 | 1,634,481 | -89 | -0.01%
Sema.RequiresOpaqueModifyCoroutineRequest | 447,561 | 447,567 | 6 | 0.0%
Sema.ResolveEffectiveMemberwiseInitRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveImplicitMemberRequest | 684,076 | 684,180 | 104 | 0.02%
Sema.ResolveProtocolNameRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeEraserTypeRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeRequest | 4,964,522 | 4,964,376 | -146 | -0.0%
Sema.ResultBuilderTypeRequest | 417,384 | 417,383 | -1 | -0.0%
Sema.ResultTypeRequest | 1,082,374 | 1,082,372 | -2 | -0.0%
Sema.RootAndResultTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.RootTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.SPIGroupsRequest | 6,630,305 | 6,632,160 | 1,854 | 0.03%
Sema.ScopedImportLookupRequest | 2,150 | 2,150 | 0 | 0.0%
Sema.SelfAccessKindRequest | 897,317 | 897,328 | 11 | 0.0%
Sema.SelfBoundsFromGenericSignatureRequest | 65,372 | 65,371 | 0 | -0.0%
Sema.SelfBoundsFromWhereClauseRequest | 1,964,491 | 1,965,573 | 1,082 | 0.06%
Sema.SetterAccessLevelRequest | 229,620 | 229,617 | -3 | -0.0%
Sema.SimpleDidSetRequest | 2,541,770 | 2,545,916 | 4,146 | 0.16%
Sema.SpecializeAttrTargetDeclRequest | 2,994 | 2,994 | 0 | 0.0%
Sema.StorageImplInfoRequest | 1,880,449 | 1,880,344 | -105 | -0.01%
Sema.StoredPropertiesAndMissingMembersRequest | 57,937 | 57,938 | 1 | 0.0%
Sema.StoredPropertiesRequest | 339,651 | 339,674 | 23 | 0.01%
Sema.StructuralRequirementsRequest | 58,032 | 58,102 | 70 | 0.12%
Sema.StructuralTypeRequest | 2,839 | 2,839 | 0 | 0.0%
Sema.SuperclassDeclRequest | 480,419 | 480,602 | 183 | 0.04%
Sema.SynthesizeAccessorRequest | 519,947 | 519,949 | 2 | 0.0%
Sema.SynthesizeDefaultInitRequest | 8,860 | 8,860 | 0 | 0.0%
Sema.SynthesizeMainFunctionRequest | 191,501 | 191,504 | 3 | 0.0%
Sema.SynthesizeMemberwiseInitRequest | 6,629 | 6,630 | 0 | 0.01%
Sema.TangentStoredPropertyRequest | 0 | 0 | 0 | 0.0%
Sema.TypeAliasRequirementsRequest | 58,032 | 58,102 | 70 | 0.12%
Sema.TypeCheckASTNodeAtLocRequest | 0 | 0 | 0 | 0.0%
Sema.TypeCheckFunctionBodyRequest | 818,056 | 818,059 | 2 | 0.0%
Sema.TypeCheckSourceFileRequest | 52,639 | 52,638 | -1 | -0.0%
Sema.TypeDeclsFromWhereClauseRequest | 52,313 | 52,315 | 1 | 0.0%
Sema.TypeEraserHasViableInitRequest | 0 | 0 | 0 | 0.0%
Sema.TypeRelationCheckRequest | 0 | 0 | 0 | 0.0%
Sema.TypeWitnessRequest | 13,625 | 13,622 | -2 | -0.02%
Sema.USRGenerationRequest | 10,709,829 | 10,727,842 | 18,013 | 0.17%
Sema.UnderlyingTypeDeclsReferencedRequest | 354,090 | 354,084 | -6 | -0.0%
Sema.UnderlyingTypeRequest | 50,039 | 50,038 | -1 | -0.0%
Sema.UnqualifiedLookupRequest | 5,882,421 | 5,882,278 | -143 | -0.0%
Sema.ValidatePrecedenceGroupRequest | 214,760 | 214,766 | 6 | 0.0%
Sema.ValueWitnessRequest | 86,785 | 86,785 | 0 | 0.0%
TBDGen.APIGenRequest | 0 | 0 | 0 | 0.0%
TBDGen.GenerateTBDRequest | 0 | 0 | 0 | 0.0%
TBDGen.PublicSymbolsRequest | 25,951 | 25,952 | 1 | 0.0%
TBDGen.SymbolSourceMapRequest | 0 | 0 | 0 | 0.0%
</details>

<a name="release-3535b3c9009a87b">

# Release

…